add: complete menu

+++ com.uet.tower-defense-common/src/main/java/com/uet/towerdefense/common/pojo/enemies/BaseEnemy.java
@@ -13,9 +13,9 @@ public interface BaseEnemy<Tid extends Object> extends BaseDynamicEntity<Tid> {
 
     void setHp(int hp);
 
-    int getSpeed();
+    double getSpeed();
 
-    void setSpeed(int speed);
+    void setSpeed(double speed);
 
     int getDefense();
